+++ src/school_supply_balance/dtos/school_supply_balance_response.rs
@@ -1,8 +1,9 @@
+use sea_orm::prelude::Decimal;
 use serde::Serialize;
 use uuid::Uuid;
 
 #[derive(Debug, Serialize)]
 pub struct SchoolSupplyBalanceResponse {
     pub id: Uuid,
-    pub balance: f64,
+    pub balance: Decimal,
 }
