+++ src/school_supply_balance/dtos/update_school_supply_balance.rs
@@ -1,7 +1,7 @@
+use sea_orm::prelude::Decimal;
 use serde::{Deserialize, Serialize};
-use uuid::Uuid;
 
 #[derive(Debug, Serialize, Deserialize)]
 pub struct UpdateSchoolSupplyBalanceDto {
-    pub balance: Option<f64>,
+    pub balance: Option<Decimal>,
 }
